AngularJS: week3 - Angular Factory, Service and Dependency Injection
How can we make use of angular services and dependency injection in order to re-factor the code to be more cleanly organized and better designed...
One of the reasons for moving the data away into the service is, eventually, that data will be moved from the service itself to a back end server.
And then, from the service, it will be able to contact the back end server, and then download the data on the fly when it is required.