AngularJS: week3 - Angular Templates, ngInclude directive
The HTML templates are purely there to specify how things should be laid out. But it is Angular that is doing all the action behind the scenes in order to render the view to the user.
So, that's why it's a template + controllers + the model, is what results in the dynamic of that application.

ngController and ngInclude directives create a new scope